#17 시퀀스 문자열 + feature/17_WorkManage_DTO,ENTITY
