+++ src/main/java/com/in4people/bootrestapi/approveWork/dto/ApplicationOvertimeWorkDTO.java
@@ -1,44 +1,31 @@
-package com.in4peopl.bootrestapi.approveWork.entity;
+package com.in4people.bootrestapi.approveWork.dto;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
 import java.util.Date;
 
-// 연장 근무 신청 Entity
-@Entity
-@Table(name = "APPLICATION_OVERTIME_WORK")
-public class ApplicationOvertimeWork {
+// 연장 근무 신청 DTO (양방향)
+public class ApplicationOvertimeWorkDTO {
 
-    @Id
-    @Column(name = "OVERTIME_NUMBER")
     private String overtimeNumber;           // 연장 근무 번호 PK
 
-    @Column(name = "OVERTIME_APPLICATION_DATE")
     private Date overtimeApplicationDate;    // 신청일
 
-    @Column(name = "OVERTIME_SCHEDULE_DATE")
     private String overtimeScheduleDate;     // 연장 근무 적용일 + 시간
 
 
     //FK
-    @Column(name = "MEM_CODE")
     private String memCode;                  // 사원번호
 
-    @Column(name = "REPORTER_TAG_CODE")
     private String reporterTagCode;          // 상신자 태그 코드
 
-    @Column(name = "WORK_SETTING_NUMBER")
     private String workSettingNumber;        // 근무 설정 번호
 
-    @Column(name = "DOCUMENT_NUMBER")
     private String documentNumber;           // 전자 결제 번호
 
-    public ApplicationOvertimeWork() {
+
+    public ApplicationOvertimeWorkDTO() {
     }
 
-    public ApplicationOvertimeWork(String overtimeNumber, Date overtimeApplicationDate, String overtimeScheduleDate, String memCode, String reporterTagCode, String workSettingNumber, String documentNumber) {
+    public ApplicationOvertimeWorkDTO(String overtimeNumber, Date overtimeApplicationDate, String overtimeScheduleDate, String memCode, String reporterTagCode, String workSettingNumber, String documentNumber) {
         this.overtimeNumber = overtimeNumber;
         this.overtimeApplicationDate = overtimeApplicationDate;
         this.overtimeScheduleDate = overtimeScheduleDate;
@@ -52,52 +39,52 @@ public class ApplicationOvertimeWork {
         return overtimeNumber;
     }
 
-    public void setOvertimeNumber(String overtimeNumber) {
-        this.overtimeNumber = overtimeNumber;
-    }
-
     public Date getOvertimeApplicationDate() {
         return overtimeApplicationDate;
-    }
-
-    public void setOvertimeApplicationDate(Date overtimeApplicationDate) {
-        this.overtimeApplicationDate = overtimeApplicationDate;
     }
 
     public String getOvertimeScheduleDate() {
         return overtimeScheduleDate;
     }
 
-    public void setOvertimeScheduleDate(String overtimeScheduleDate) {
-        this.overtimeScheduleDate = overtimeScheduleDate;
-    }
-
     public String getMemCode() {
         return memCode;
-    }
-
-    public void setMemCode(String memCode) {
-        this.memCode = memCode;
     }
 
     public String getReporterTagCode() {
         return reporterTagCode;
     }
 
-    public void setReporterTagCode(String reporterTagCode) {
-        this.reporterTagCode = reporterTagCode;
-    }
-
     public String getWorkSettingNumber() {
         return workSettingNumber;
     }
 
-    public void setWorkSettingNumber(String workSettingNumber) {
-        this.workSettingNumber = workSettingNumber;
-    }
-
     public String getDocumentNumber() {
         return documentNumber;
+    }
+
+    public void setOvertimeNumber(String overtimeNumber) {
+        this.overtimeNumber = overtimeNumber;
+    }
+
+    public void setOvertimeApplicationDate(Date overtimeApplicationDate) {
+        this.overtimeApplicationDate = overtimeApplicationDate;
+    }
+
+    public void setOvertimeScheduleDate(String overtimeScheduleDate) {
+        this.overtimeScheduleDate = overtimeScheduleDate;
+    }
+
+    public void setMemCode(String memCode) {
+        this.memCode = memCode;
+    }
+
+    public void setReporterTagCode(String reporterTagCode) {
+        this.reporterTagCode = reporterTagCode;
+    }
+
+    public void setWorkSettingNumber(String workSettingNumber) {
+        this.workSettingNumber = workSettingNumber;
     }
 
     public void setDocumentNumber(String documentNumber) {
@@ -106,7 +93,7 @@ public class ApplicationOvertimeWork {
 
     @Override
     public String toString() {
-        return "ApplicationOvertimeWork{" +
+        return "ApplicationOvertimeWorkDTO{" +
                 "overtimeNumber='" + overtimeNumber + '\'' +
                 ", overtimeApplicationDate=" + overtimeApplicationDate +
                 ", overtimeScheduleDate='" + overtimeScheduleDate + '\'' +
@@ -117,5 +104,3 @@ public class ApplicationOvertimeWork {
                 '}';
     }
 }
-
-
