+++ src/main/java/com/in4people/bootrestapi/approveWork/entity/AlldayCategory.java
@@ -1,4 +1,4 @@
-package com.in4peopl.bootrestapi.approveWork.entity;
+package com.in4people.bootrestapi.approveWork.entity;
 
 // 휴가 전일/반일 entity
