+++ src/main/java/com/in4people/bootrestapi/approveWork/entity/DocumentApproval.java
@@ -1,4 +1,4 @@
-package com.in4peopl.bootrestapi.approveWork.entity;
+package com.in4people.bootrestapi.approveWork.entity;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
